アップロード完成

+++ backend/internal/dto/photo_dto.go
@@ -2,5 +2,5 @@ package dto
 
 // PhotoUploadResponse は画像アップロード後に返すID配列
 type PhotoUploadResponse struct {
-	PhotoIDs []uint `json:"photo_ids"`
+	PhotoIDs []int64 `json:"photo_ids"`
 }
